+++ tests/Application/ProductUseCasesFacade.php
@@ -10,7 +10,7 @@ use WonderNetwork\SlimKernelTestingHarness\UseCase\KernelHttpClientUseCase;
 final readonly class ProductUseCasesFacade {
     use KernelHttpClientUseCase;
 
-    public function add(string $name, float $price): HttpResponseAssertion {
+    public function add(string $name, int $price): HttpResponseAssertion {
         // shh, compact was not safe in the PHP 4 days, but it’s
         // perfectly safe now, once we have all this static analysis
         // watching over our backs
@@ -23,7 +23,7 @@ final readonly class ProductUseCasesFacade {
         );
     }
 
-    public function update(string $id, string $name = null, float $price = null): HttpResponseAssertion {
+    public function update(string $id, string $name = null, int $price = null): HttpResponseAssertion {
         return $this->httpClient->patch("/product/$id", compact('name', 'price'));
     }
 }
